Fixed a bug with loading guild details on zoning which would cause guild permissions to be out of sync

+++ zone/guild_mgr.cpp
@@ -425,15 +425,15 @@ bool ZoneGuildManager::VerifyAndClearInvite(uint32 char_id, uint32 guild_id, uin
 
 void ZoneGuildManager::ProcessWorldPacket(ServerPacket *pack) 
 {
-	switch (pack->opcode) 
+	switch (pack->opcode)
 	{
-	case ServerOP_RefreshGuild: 
+	case ServerOP_RefreshGuild:
 	{
 		if (pack->size != sizeof(ServerGuildRefresh_Struct)) {
 			LogError("Received ServerOP_RefreshGuild of incorrect size [{}], expected [{}]", pack->size, sizeof(ServerGuildRefresh_Struct));
 			return;
 		}
-	
+
 		ServerGuildRefresh_Struct* s = (ServerGuildRefresh_Struct*)pack->pBuffer;
 		LogDebug("Received guild refresh from world for guild id [{}] changes: name=[{}] motd=[{}] rank=[{}] relation=[{}]",
 			s->guild_id,
@@ -444,7 +444,7 @@ void ZoneGuildManager::ProcessWorldPacket(ServerPacket *pack)
 		);
 
 		RefreshGuild(s->guild_id);
-		
+
 		if (s->motd_change) {
 			//resend guild MOTD to all guild members in this zone.
 			entity_list.SendGuildMOTD(s->guild_id);
@@ -467,7 +467,7 @@ void ZoneGuildManager::ProcessWorldPacket(ServerPacket *pack)
 		break;
 	}
 
-	case ServerOP_GuildCharRefresh: 
+	case ServerOP_GuildCharRefresh:
 	{
 		if (pack->size != sizeof(ServerGuildCharRefresh_Struct)) {
 			LogError("Received ServerOP_RefreshGuild of incorrect size [{}], expected [{}]", pack->size, sizeof(ServerGuildCharRefresh_Struct));
@@ -507,7 +507,6 @@ void ZoneGuildManager::ProcessWorldPacket(ServerPacket *pack)
 			if (c->ClientVersion() >= EQ::versions::ClientVersion::RoF)
 			{
 				c->SendGuildRanks();
-//				entity_list.SendToGuildTitleDisplay(c);
 			}
 		}
 
@@ -518,7 +517,6 @@ void ZoneGuildManager::ProcessWorldPacket(ServerPacket *pack)
 	case ServerOP_GuildRankUpdate:
 	{
 		ServerGuildRankUpdate_Struct* sgrus = (ServerGuildRankUpdate_Struct*)pack->pBuffer;
-		RefreshGuild(sgrus->GuildID);
 
 		if (is_zone_loaded)
 		{
@@ -538,7 +536,7 @@ void ZoneGuildManager::ProcessWorldPacket(ServerPacket *pack)
 			strn0cpy(gsrs->MemberName, sgrus->MemberName, sizeof(gsrs->MemberName));
 			gsrs->Banker = sgrus->Banker;
 
-			entity_list.QueueClientsGuild(nullptr, outapp, false, sgrus->GuildID);
+			entity_list.QueueClientsGuild(outapp, sgrus->GuildID);
 
 			safe_delete(outapp);
 		}
@@ -552,28 +550,31 @@ void ZoneGuildManager::ProcessWorldPacket(ServerPacket *pack)
 		break;
 	}
 
-	case ServerOP_DeleteGuild: 
+	case ServerOP_DeleteGuild:
 	{
 		if (pack->size != sizeof(ServerGuildID_Struct)) {
 			LogError("Received ServerOP_DeleteGuild of incorrect size [{}], expected [{}]", pack->size, sizeof(ServerGuildID_Struct));
 			return;
 		}
-	
-		ServerGuildID_Struct* s = (ServerGuildID_Struct*)pack->pBuffer;
 
-		LogGuilds("Received guild delete from world for guild [{}]", s->guild_id);
-		
-		auto clients = entity_list.GetClientList();
-		for (auto& c : clients) {
-			if (c.second->GuildID() == s->guild_id) {
-				c.second->RefreshGuildInfo();
-				c.second->SendGuildMembers();
-				c.second->MessageString(Chat::Guild, GUILD_DISBANDED);
+		if (is_zone_loaded)
+		{
+			ServerGuildID_Struct* s = (ServerGuildID_Struct*)pack->pBuffer;
+
+			LogGuilds("Received guild delete from world for guild [{}]", s->guild_id);
+
+			auto clients = entity_list.GetClientList();
+			for (auto& c : clients) {
+				if (c.second->GuildID() == s->guild_id) {
+					c.second->RefreshGuildInfo();
+					c.second->SendGuildMembers();
+					c.second->MessageString(Chat::Guild, GUILD_DISBANDED);
+				}
 			}
-		}
 
-		LoadGuilds();
-		break;
+			LoadGuilds();
+			break;
+		}
 	}
 
 	case ServerOP_GuildMemberUpdate:
@@ -585,7 +586,6 @@ void ZoneGuildManager::ProcessWorldPacket(ServerPacket *pack)
 			auto outapp = new EQApplicationPacket(OP_GuildMemberUpdate, sizeof(GuildMemberUpdate_Struct));
 
 			GuildMemberUpdate_Struct* gmus = (GuildMemberUpdate_Struct*)outapp->pBuffer;
-			RefreshGuild(gmus->GuildID);
 
 			gmus->GuildID = sgmus->GuildID;
 			strn0cpy(gmus->MemberName, sgmus->MemberName, sizeof(gmus->MemberName));
@@ -593,7 +593,7 @@ void ZoneGuildManager::ProcessWorldPacket(ServerPacket *pack)
 			gmus->InstanceID = 0;	// I don't think we care what Instance they are in, for the Guild Management Window.
 			gmus->LastSeen = sgmus->LastSeen;
 
-			entity_list.QueueClientsGuild(nullptr, outapp, false, sgmus->GuildID);
+			entity_list.QueueClientsGuild(outapp, sgmus->GuildID);
 
 			safe_delete(outapp);
 		}
@@ -669,7 +669,7 @@ void ZoneGuildManager::ProcessWorldPacket(ServerPacket *pack)
 			gts->Toggle = Toggle;
 			gts->TimePosted = TimePosted;
 			gts->Name[0] = 0;
-			entity_list.QueueClientsGuild(nullptr, outapp, false, GuildID);
+			entity_list.QueueClientsGuild(outapp, GuildID);
 			safe_delete(outapp);
 			break;
 		}
@@ -679,7 +679,14 @@ void ZoneGuildManager::ProcessWorldPacket(ServerPacket *pack)
 		if (is_zone_loaded)
 		{
 			ServerGuildPermissionUpdate_Struct* sgpus = (ServerGuildPermissionUpdate_Struct*)pack->pBuffer;
-			RefreshGuild(sgpus->GuildID);
+			auto res = m_guilds.find(sgpus->GuildID);
+			if (sgpus->FunctionValue) {
+				res->second->functions[sgpus->FunctionID].perm_value |= (1UL << (8 - sgpus->Rank));
+			}
+			else {
+				res->second->functions[sgpus->FunctionID].perm_value &= ~(1UL << (8 - sgpus->Rank));
+			}
+
 			auto client = entity_list.GetMob(sgpus->MemberName);
 			auto outapp = new EQApplicationPacket(OP_GuildUpdateURLAndChannel, sizeof(GuildPermission_Struct));
 			GuildPermission_Struct* guuacs = (GuildPermission_Struct*)outapp->pBuffer;
@@ -688,8 +695,8 @@ void ZoneGuildManager::ProcessWorldPacket(ServerPacket *pack)
 			guuacs->function_id = sgpus->FunctionID;
 			guuacs->value = sgpus->FunctionValue;
 
-			entity_list.QueueClientsGuild(client, outapp, true, sgpus->GuildID);
-			delete(outapp);
+			entity_list.QueueClientsGuild(outapp, sgpus->GuildID);
+			safe_delete(outapp);
 
 			LogDebug("Zone Received guild permission update from world for rank {} function id [{}] and value [{}]",
 				guuacs->rank = sgpus->Rank,
@@ -704,19 +711,23 @@ void ZoneGuildManager::ProcessWorldPacket(ServerPacket *pack)
 	}
 	case ServerOP_GuildRankNameChange:
 	{
-		ServerGuildRankNameChange* s = (ServerGuildRankNameChange*)pack->pBuffer;
-		LogGuilds("Received guild rank name change from world for rank [{}] from guild [{}]", s->rank, s->guild_id);
-		RefreshGuild(s->guild_id);
-		auto outapp = new EQApplicationPacket(OP_GuildUpdateURLAndChannel, sizeof(GuildUpdateUCP));
-		GuildUpdateUCP* gucp = (GuildUpdateUCP*)outapp->pBuffer;
-		gucp->payload.rank_name.rank = s->rank;
-		strcpy(gucp->payload.rank_name.rank_name, s->rank_name);
-		gucp->action = 4;
+		if (is_zone_loaded)
+		{
+			ServerGuildRankNameChange* s = (ServerGuildRankNameChange*)pack->pBuffer;
+			LogGuilds("Received guild rank name change from world for rank [{}] from guild [{}]", s->rank, s->guild_id);
 
-		entity_list.QueueClientsGuild(nullptr, outapp, false, s->guild_id);
+			auto guild = GetGuildByGuildID(s->guild_id);
+			auto outapp = new EQApplicationPacket(OP_GuildUpdateURLAndChannel, sizeof(GuildUpdateUCP));
+			GuildUpdateUCP* gucp = (GuildUpdateUCP*)outapp->pBuffer;
+			gucp->payload.rank_name.rank = s->rank;
+			strcpy(gucp->payload.rank_name.rank_name, s->rank_name.c_str());
+			gucp->action = 4;
 
-		safe_delete(outapp);
-		break;
+			entity_list.QueueClientsGuild(outapp, s->guild_id);
+
+			safe_delete(outapp);
+			break;
+		}
 	}
 	}
 }
@@ -1553,8 +1564,8 @@ bool GuildBankManager::AllowedToWithdraw(uint32 GuildID, uint16 Area, uint16 Slo
 	return false;
 }
 
-void ZoneGuildManager::UpdateRankPermission(uint32 gid, uint32 charid, uint32 fid, uint32 rank, uint32 value) {
-
+void ZoneGuildManager::UpdateRankPermission(uint32 gid, uint32 charid, uint32 fid, uint32 rank, uint32 value) 
+{
 	auto res = m_guilds.find(gid);
 	if (value) {
 		res->second->functions[fid].perm_value |= (1UL << (8 - rank));
@@ -1587,18 +1598,18 @@ void ZoneGuildManager::UpdateRankName(uint32 guild_id, uint32 rank, std::string 
 	out.guild_id = guild_id;
 	out.rank = rank;
 	out.title = rank_name;
-	GuildRanksRepository::UpdateOne(*m_db, out);
+	GuildRanksRepository::UpdateTitle(*m_db, out);
 }
 
-void ZoneGuildManager::SendRankName(uint32 guild_id, uint32 rank, std::string rank_name)
+void ZoneGuildManager::SendRankName(uint32 guild_id, uint32 rank, std::string& rank_name)
 {
 	auto pack = new ServerPacket(ServerOP_GuildRankNameChange, sizeof(ServerGuildRankNameChange));
 	ServerGuildRankNameChange* sgpus = (ServerGuildRankNameChange*)pack->pBuffer;
 
 	sgpus->guild_id = guild_id;
 	sgpus->rank = rank;
-	strcpy(sgpus->rank_name, rank_name.c_str());
-	worldserver.SendPacket(pack);
+	sgpus->rank_name = static_cast<std::string>(rank_name);
+	worldserver.SendPacket(pack);	
 	safe_delete(pack);
 }
